Change submission status to an enum
This has multiple advantages: K2 can start making a difference between
submissions that are wrong and those that might be correct, because
their results are not shown on the scoreboard. This method is also
extendable in the future.

The migration is slightly lossy, which isn't too big of a deal. When the
renewed submission correctness checker is written in the near future, you
should reprocess all submissions. It could be done at this moment as well,
but that would be a waste of cycles.