Add new version of job correctness checker.
This properly handles overcomplete submissions, which are made after the
first correct submission. This will mark them as "overcomplete" instead
of wrong (if we already have an correct submission) or correct (when we
recreate the database at a later point). It will also nicely mark
submissions against hidden problems or made with hidden accounts.

Known issue: when submissions are made in the same minute, both will be
marked as correct, where only one would be more desireable.