Add optional Sentry integration for error reporting in production
Logs generated by K2 tend to be huge and easily spammed over by giant
stacktraces (Ruby stacktraces are usually around 90 frames deep) and
digging through this log isn't fun. Inspired my Mike Perham's, the
developer of Sidekiq, blog post [1] I decided to investigate into Error
Services. I'm personally opposed to using SaaS, because it is often unclear
what the software is doing. I'm making an exception for Sentry for the
following reasons:

- The Sentry server is open source under the 3-clause BSD license with
  instructions provided to set up an self-hosted instance in the future.
  The client code is licensed under the Apache license. If Sentry proves
  to be useful, I'm planning to move Sentry to a self-hosted instance.
- Sentry is implemented in such a way that it is disabled by default and
  can be enabled by adding a single line in the secrets file. Old
  releases of K2 can therefore still be used without implications for
  the user functionality, even after Sentry has been discontinued.
- The data saved by Sentry is only relevant for a few days, when the bug
  is fixed, this data is no longer relevant. Therefore Sentry is not
  holding valuable data hostage.
- Data saved by Sentry can be considered slightly privacy intrusive: it
  contains user agents, IP addresses, which is in theory enough to identify
  users. This information is exposed by the client when connecting to
  the site. This information might be useful for detecting issues
  that only appear in certain places or with certain browsers. The error
  report also contains various information about the code, but this is
  considered public knowledge anyway, since the code is on GitHub.

With that said, default behavior is that errors are not reported, but
this can be easily enabled by adding an atrribute "sentry_dsn" to the
secrets.yml file. I hope that with the Sentry integration, bugs can be
fixed quickly and efficiently.

[1]: http://www.mikeperham.com/2013/08/25/please-use-an-error-service/