Add tooltip to Submissions to clarify status, prevent possible crash
A tooltip was added to the submissions tab to clarify the status of
submissions, because in some cases this was not directly obvious, like
when a submission is overcomplete.

The crash is related to problems where the problem was hidden from the
server. This is a very shallow fix where the submission is hidden when
there is no problem present. The underlying issue is that there are now
submissions that are not linked to a server (since the problem determines
the linked server) :/