Initial version of Webhook support
Currently there is only one webhook that sends out a very limited amount
of data. The initial use case is a slack bot that sends messages when
submissions are marked correct. If someone wants to have extra data
added to help with their application or an extra webhook, I'm happy to help.